Add files via upload
These are the files I made to start this project. The goal is to create a simple, sleek and efficient application for managing new and used tires, buying and selling of tires and the retreading of tires. Program should be able to catch and alert for discrepancies in inventory and input into databases between the warehouse and service sides company using this system. Some customers have special accounts, special requests and all circumstances of business will not be so cut and dry... the program should be able to handle this type of concern while still remaining sleek and user friendly.